Renamed ConstantExpression to LiteralExpression and refactored binary and unary expression functions
