+++ StepLang.Tests/Framework/Mutating/DoRemoveFunctionTest.cs
@@ -31,12 +31,12 @@ public class DoRemoveFunctionTest
     {
         public DoRemoveData()
         {
-            Add(ListResult.Empty, ConstantExpression.Str("test"), ListResult.Empty);
-            Add(ListResult.Empty, ConstantExpression.Number(1), ListResult.Empty);
-            Add(ListResult.Empty, ConstantExpression.True, ListResult.Empty);
-            Add(ListResult.Empty, Expression.Add(ConstantExpression.Number(1), ConstantExpression.Number(2)), ListResult.Empty);
-            Add(ListResult.From(BoolResult.True), ConstantExpression.False, ListResult.From(BoolResult.True));
-            Add(ListResult.From(BoolResult.False), ConstantExpression.False, ListResult.Empty);
+            Add(ListResult.Empty, LiteralExpression.Str("test"), ListResult.Empty);
+            Add(ListResult.Empty, LiteralExpression.Number(1), ListResult.Empty);
+            Add(ListResult.Empty, LiteralExpression.True, ListResult.Empty);
+            Add(ListResult.Empty, BinaryExpression.Add(LiteralExpression.Number(1), LiteralExpression.Number(2)), ListResult.Empty);
+            Add(ListResult.From(BoolResult.True), LiteralExpression.False, ListResult.From(BoolResult.True));
+            Add(ListResult.From(BoolResult.False), LiteralExpression.False, ListResult.Empty);
         }
     }
 }
