+++ StepLang/Expressions/Results/ExpressionResult.cs
@@ -121,5 +121,5 @@ public abstract class ExpressionResult : IEquatable<ExpressionResult>
 
     public abstract ExpressionResult DeepClone();
 
-    public ConstantExpression ToExpression() => new(this);
+    public LiteralExpression ToLiteralExpression() => LiteralExpression.From(this);
 }
