+++ StepLang/Expressions/Results/FunctionResult.cs
@@ -5,7 +5,7 @@ namespace StepLang.Expressions.Results;
 
 public class FunctionResult : ValueExpressionResult<FunctionDefinition>
 {
-    public static FunctionResult VoidFunction => new UserDefinedFunctionDefinition(new List<(Token type, Token identifier)>(), new[] { new ReturnStatement(ConstantExpression.Void) }).ToResult();
+    public static FunctionResult VoidFunction => new UserDefinedFunctionDefinition(new List<(Token type, Token identifier)>(), Array.Empty<Statement>()).ToResult();
 
     /// <inheritdoc />
     public FunctionResult(FunctionDefinition value) : base(ResultType.Function, value)
